refactor(typescript): split the extractors, and drop what never ran
The same interleaving as the Rust side: two unrelated types in one file. Four
modules now each answer one question — what a file declares, what a pattern binds,
what an import introduces, what a file reads.

Four functions went rather than moved. `extract_member_usage`, `extract_new_usage`,
`extract_parameter_definition` and `determine_typescript_accessibility` were 89
lines that nothing called, each carrying an `#[allow(dead_code)]` to silence the
compiler pointing that out. The file now has none.

Behaviour is unchanged: 743 of 743, precision and recall 1.000, identical to the
recorded baseline.
